refactor: split DirectorService facade into focused director modules (#10)
* refactor(director): extract runtime helpers to runtime.ts (Step 2/9)

Move 5 runtime/orchestration helpers (scheduleBackgroundRun,
withWorkflowTaskUsage, markTaskFailedWithLatestDiagnostic,
markDirectorTaskRunning, assertHighMemoryDirectorStartAllowed) plus
2 utility functions (isWorkflowTaskCancelledError, stringifyDirectorTaskError)
from DirectorService.ts to runtime.ts.

Instance methods on NovelDirectorService stay as one-line delegators so
tests that monkey-patch them (novelDirectorRetry.test.js) keep working.

Also: gitignore .playwright-cli/ artifacts.

Tests: 643 / 641 pass / 2 skipped (matches baseline).

* refactor(director): extract style-context helpers to styleContext.ts (Step 3/9)

Move enrichDirectorStyleContext and ensurePrimaryNovelStyleBinding
to styleContext.ts. Instance methods stay as one-line delegators.

Tests: 643 / 641 pass / 2 skipped (matches baseline).

* refactor(director): extract phase resolver helpers to phaseResolver.ts (Step 4/9)

Move 6 phase/resume resolver helpers to phaseResolver.ts.
Instance methods stay as one-line delegators so existing tests can monkey-patch them.

Tests: 643 / 641 pass / 2 skipped (matches baseline).

* refactor(director): extract candidate stage helpers to candidateStageRuntime.ts (Step 5/9)

Move candidate-stage detection, request shaping, resume scheduling, and failure handling into candidateStageRuntime.ts.
Instance methods stay as wrappers so retry tests can keep monkey-patching continueCandidateStageTask.

Tests: 643 / 641 pass / 2 skipped (matches baseline).

* refactor(director): extract pipeline runner to pipelineRunner.ts (Step 6/9)

Move pipeline phase orchestration and auto-approval progression into pipelineRunner.ts.
Instance methods stay as wrappers so retry tests can keep monkey-patching runDirectorPipeline.

Tests: 643 / 641 pass / 2 skipped (matches baseline).

* refactor(director): finish splitting director facade flows (Step 7-10/9)

Extract confirm-candidate, continue-task, repair-title, takeover, seed-payload, and phase-execution helpers out of DirectorService.
Keep service wrappers and this.method(...) call sites so existing monkey-patched tests continue to work.
Remove DirectorService.ts from source-length allowlist after reducing it below 500 lines.

Checks: pnpm typecheck; pnpm lint; pnpm check:source-length; pnpm --filter @ai-novel/server test (643 / 641 pass / 2 skipped).
