+++ scripts/source-file-length.allowlist.json
@@ -11,7 +11,6 @@
   "server/src/services/character/CharacterLibrarySyncService.ts": 724,
   "server/src/services/novel/NovelExportService.ts": 772,
   "server/src/services/novel/characterPrep/CharacterPreparationService.ts": 722,
-  "server/src/services/director/DirectorService.ts": 1682,
   "server/src/services/director/takeover.ts": 1023,
   "server/src/services/novel/dynamics/CharacterDynamicsMutationService.ts": 716,
   "server/src/services/novel/novelCorePipelineService.ts": 827,
